backend/WEB-AUDIT-BE/requirement.txt
flask
flask-cors
flask-session
pyjwt
semgrep
flask_dance
requests